Extract single_color into its own file
