+++ apa102/single_color.h
@@ -1,0 +1,8 @@
+#ifndef SINGLE_COLOR_H
+#define SINGLE_COLOR_H
+
+int singlecolor_main(struct apa102_led color);
+int rgb_main(float r, float g, float b);
+int hsv_main(float h, float s, float v);
+
+#endif
